added Project logic and refactored

+++ server/src/main/java/com/taskflow/taskflow/service/TeamServiceImpl.java
@@ -25,12 +25,20 @@ public class TeamServiceImpl implements TeamService {
     private TeamRepository teamRepository;
     private UserRepository userRepository;
     private TeamMemberRepository teamMemberRepository;
+    private TeamAccessService teamAccessService;
+    private AuthService authService;
 
     @Autowired
-    public TeamServiceImpl(TeamRepository teamRepository, UserRepository userRepository, TeamMemberRepository teamMemberRepository) {
+    public TeamServiceImpl(TeamRepository teamRepository,
+                           UserRepository userRepository,
+                           TeamMemberRepository teamMemberRepository,
+                           TeamAccessService teamAccessService,
+                           AuthService authService) {
         this.teamRepository = teamRepository;
         this.userRepository = userRepository;
         this.teamMemberRepository = teamMemberRepository;
+        this.teamAccessService = teamAccessService;
+        this.authService = authService;
     }
 
     @Override
@@ -80,17 +88,21 @@ public class TeamServiceImpl implements TeamService {
 
     // Need to ensure User is validated and team owner
     @Override
-    public Team updateTeam(User currentUser, int teamId, UpdateTeamRequest request) {
-        Team team = teamRepository.findById(teamId).orElseThrow(() -> new ResourceNotFoundException("Team not found"));
+    public Team updateTeam(int teamId, UpdateTeamRequest request) {
+            User currentUser = authService.getCurrentUser();
 
-        // Check requesting user is an OWNER of this team
-        TeamMember membership = teamMemberRepository
-                .findByTeamIdAndUserId(teamId, currentUser.getId())
-                .orElseThrow(() -> new AccessDeniedException("Not a member of this team"));
+            Team team = teamRepository.findById(teamId).orElseThrow(() -> new ResourceNotFoundException("Team not found"));
+//
+//        // Check requesting user is an OWNER of this team
+//        TeamMember membership = teamMemberRepository
+//                .findByTeamIdAndUserId(teamId, currentUser.getId())
+//                .orElseThrow(() -> new AccessDeniedException("Not a member of this team"));
 
-        if (membership.getRole() != TeamRole.OWNER) {
-            throw new AccessDeniedException("Only team owners can update the team");
-        }
+        //teamAccessService.validateTeamAccess(teamId, currentUser.getId());
+        teamAccessService.validateOwnerAccess(teamId, currentUser.getId());
+//        if (membership.getRole() != TeamRole.OWNER) {
+//            throw new AccessDeniedException("Only team owners can update the team");
+//        }
 
         // update fields only if present
         if(request.getName() != null) team.setName(request.getName());
@@ -104,9 +116,12 @@ public class TeamServiceImpl implements TeamService {
 
     @Override
     public void deleteById(int id) {
+        User currentUser = authService.getCurrentUser();
+
         // ensure team exists
         Team team = teamRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Team not found: " + id));
+        teamAccessService.validateOwnerAccess(id, currentUser.getId());
 
         teamRepository.delete(team);
     }
@@ -117,33 +132,41 @@ public class TeamServiceImpl implements TeamService {
 
     @Override
     public List<TeamMember> getTeamMembers(
-            int teamId,
-            User currentUser
-    ) {
-        teamRepository.findById(teamId).orElseThrow(() -> new ResourceNotFoundException("Team not found"));
+            int teamId) {
+        User currentUser = authService.getCurrentUser();
+        teamAccessService.validateOwnerAccess(teamId, currentUser.getId());
 
-        // Must be a team member or owner to view the list of members
-        if (!teamMemberRepository.existsByTeamIdAndUserId(teamId, currentUser.getId())) {
-            throw new AccessDeniedException("Not a member of this team");
-        }
+//        teamRepository.findById(teamId).orElseThrow(() -> new ResourceNotFoundException("Team not found"));
+//
+//        // Must be a team member or owner to view the list of members
+//        if (!teamMemberRepository.existsByTeamIdAndUserId(teamId, currentUser.getId())) {
+//            throw new AccessDeniedException("Not a member of this team");
+//        }
 
         return teamMemberRepository.findByTeamId(teamId);
     }
 
     // NOTE ~ On the client side will need a way to handle this to avoid multiple requests
     @Override
-    public void addTeamMembers(int teamId, ManageTeamMemberRequest request, User currentUser) {
+    public void addTeamMembers(int teamId, ManageTeamMemberRequest request) {
+//        Team team = teamRepository.findById(teamId)
+//                .orElseThrow(() -> new ResourceNotFoundException("Team not found"));
+//
+//        // Only OWNER can add members
+//        TeamMember membership = teamMemberRepository
+//                .findByTeamIdAndUserId(teamId, currentUser.getId())
+//                .orElseThrow(() -> new AccessDeniedException("Not a member of this team"));
+//
+//        if (membership.getRole() != TeamRole.OWNER) {
+//            throw new AccessDeniedException("Only owners can add members");
+//        }
+
+        User currentUser = authService.getCurrentUser();
+        // ensure user has correct team acccess
+        teamAccessService.validateOwnerAccess(teamId, currentUser.getId());
         Team team = teamRepository.findById(teamId)
                 .orElseThrow(() -> new ResourceNotFoundException("Team not found"));
 
-        // Only OWNER can add members
-        TeamMember membership = teamMemberRepository
-                .findByTeamIdAndUserId(teamId, currentUser.getId())
-                .orElseThrow(() -> new AccessDeniedException("Not a member of this team"));
-
-        if (membership.getRole() != TeamRole.OWNER) {
-            throw new AccessDeniedException("Only owners can add members");
-        }
 
         // Add each userId to team
         // Wow this is clean
@@ -165,19 +188,12 @@ public class TeamServiceImpl implements TeamService {
 
     // remove team members, will use same pattern as passing in User IDs in an array to avoid
     // multiple API calls
+    @Override
     @Transactional
-    public void removeTeamMembers(int teamId, ManageTeamMemberRequest request, User currentUser) {
-        Team team = teamRepository.findById(teamId)
-                .orElseThrow(() -> new ResourceNotFoundException("Team not found"));
-
-        // Only OWNER can add members
-        TeamMember membership = teamMemberRepository
-                .findByTeamIdAndUserId(teamId, currentUser.getId())
-                .orElseThrow(() -> new AccessDeniedException("Not a member of this team"));
-
-        if (membership.getRole() != TeamRole.OWNER) {
-            throw new AccessDeniedException("Only owners can add members");
-        }
+    public void removeTeamMembers(int teamId, ManageTeamMemberRequest request) {
+        User currentUser = authService.getCurrentUser();
+        // ensure user has correct team acccess
+        teamAccessService.validateOwnerAccess(teamId, currentUser.getId());
 
         // Prevent owner from removing themselves
         if (request.getUserIds().contains(currentUser.getId())) {
